pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "product-builders"
version = "0.1.0"
description = "Analyze product codebases and generate tailored Cursor rules, governance, and onboarding guides"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11,<4.0"
authors = [
    {name = "Platform Team"},
]
dependencies = [
    "click>=8.1",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "rich>=13.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]
webapp = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
]

[project.scripts]
product-builders = "product_builders.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
